requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.2
scikit-learn>=0.24.2
scikit-image>=0.18.1
opencv-python>=4.5.2
tqdm>=4.61.1
lime>=0.2.0.1
shap>=0.39.0
albumentations>=1.0.3
tensorboard>=2.6.0
pytest>=6.2.5
Pillow>=8.2.0
jupyterlab>=3.0.16
seaborn>=0.11.1
flake8>=3.9.2
black>=21.5b2
isort>=5.9.1
kaggle>=1.5.12
